+++ src/cpp/benders/benders_by_batch/include/BendersByBatch.h
@@ -9,8 +9,8 @@ class BendersByBatch : public BendersMpi {
 
  public:
   BendersByBatch(BendersBaseOptions const &options, Logger logger,
-                 Writer writer, mpi::environment &env,
-                 mpi::communicator &world);
+                 Writer writer, mpi::environment &env, mpi::communicator &world,
+                 std::shared_ptr<MathLoggerDriver> mathLoggerDriver);
   ~BendersByBatch() override = default;
   void Run() override;
   void BuildCut(const std::vector<std::string> &batch_sub_problems,
