+++ src/cpp/benders/benders_core/CMakeLists.txt
@@ -26,7 +26,9 @@ add_library (benders_core STATIC
 	${CMAKE_CURRENT_SOURCE_DIR}/LastIterationWriter.cpp 
 	${CMAKE_CURRENT_SOURCE_DIR}/LastIterationReader.cpp 
 	${CMAKE_CURRENT_SOURCE_DIR}/LastIterationPrinter.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/StartUp.cpp)
+	${CMAKE_CURRENT_SOURCE_DIR}/StartUp.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/BendersMathLogger.cpp
+	)
 
 get_target_property(JSON_INC_PATH jsoncpp_lib INTERFACE_INCLUDE_DIRECTORIES)
