+++ src/cpp/benders/logger/MathLogger.cpp
@@ -71,15 +71,3 @@ MathLoggerFile::MathLoggerFile(const std::filesystem::path &filename)
               << "Invalid file name passed as parameter" << std::endl;
   }
 }
-
-void MathLoggerDriver::write_header() {
-  for (auto logger : math_loggers_) {
-    logger->write_header();
-  }
-}
-
-void MathLoggerDriver::add_logger(MathLogger* logger) {
-  if (logger) {
-    math_loggers_.push_back(logger);
-  }
-}
